+++ src/webots_staging/render/capability/mod.rs
@@ -6,8 +6,8 @@ pub mod vision;
 use webots_proto::r2025a::Node;
 
 use crate::webots_staging::scene::WebotsSceneDescription;
-use phoxal_component::v1::capability::Capability as PhysicalCapability;
-use phoxal_simulation::capability::Capability as SimulationCapability;
+use phoxal_core_component::v1::capability::Capability as PhysicalCapability;
+use phoxal_core_simulation::capability::Capability as SimulationCapability;
 
 impl WebotsSceneDescription {
     pub fn render_link_capability(